discordBackend/pyproject.toml
[project]
name = "discordbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0.0",
    "discord.py>=2.5.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "structlog>=25.4.0",
    "uvicorn>=0.34.3",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
]
